HAWKULAR-217 : Initial implementation for JVM Metrics tab. Includes other (needed) changes
Included changes for dealing better with tabs and splitting the Details controllers
