+++ ui/console/src/main/scripts/plugins/metrics/plugins/metrics/html/app-details/detail-datasources.html
@@ -1,41 +1,40 @@
+<section id="hk-datasources" class="hk-tab-content">
 
-<section id="hk-datasources" class="hk-tab-content"  ng-if="vm.activeTab === 'datasources'">
-
-    <span class="hk-update pull-right" tooltip-trigger tooltip-placement="bottom" tooltip="Last update"><i class="fa fa-clock-o"></i>{{vm.endTimeStamp | date:'mediumTime'}}</span>
-    <h3 class="pull-left">Fake Datasource 1</h3>
-    <div class="panel panel-default hk-summary">
-        <div class="row">
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">2 <i class="fa fa-flag"></i></span>
-                <span class="hk-item">Datasource Alerts</span>
-            </div>
-            <div class="col-sm-6 hk-summary-item">
-                <span class="hk-data">Chart TBD</span>
-                <span class="hk-item">Connections Used %</span>
-            </div>
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">7 Available (of 28)</span>
-                <span class="hk-item">Available Connections (of Total)</span>
-            </div>
-        </div>
+  <span class="hk-update pull-right" tooltip-trigger tooltip-placement="bottom" tooltip="Last update"><i class="fa fa-clock-o"></i>{{vm.endTimeStamp | date:'mediumTime'}}</span>
+  <h3 class="pull-left">Fake Datasource 1</h3>
+  <div class="panel panel-default hk-summary">
+    <div class="row">
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data">2 <i class="fa fa-flag"></i></span>
+        <span class="hk-item">Datasource Alerts</span>
+      </div>
+      <div class="col-sm-6 hk-summary-item">
+        <span class="hk-data">Chart TBD</span>
+        <span class="hk-item">Connections Used %</span>
+      </div>
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data">7 Available (of 28)</span>
+        <span class="hk-item">Available Connections (of Total)</span>
+      </div>
     </div>
+  </div>
 
-    <h3 class="pull-left">Fake Datasource 2</h3>
-    <div class="panel panel-default hk-summary">
-        <div class="row">
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">2 <i class="fa fa-flag"></i></span>
-                <span class="hk-item">Datasource Alerts</span>
-            </div>
-            <div class="col-sm-6 hk-summary-item">
-                <span class="hk-data">Chart TBD</span>
-                <span class="hk-item">Connections Used %</span>
-            </div>
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">45 Available (of 50)</span>
-                <span class="hk-item">Available Connections (of Total)</span>
-            </div>
-        </div>
+  <h3 class="pull-left">Fake Datasource 2</h3>
+  <div class="panel panel-default hk-summary">
+    <div class="row">
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data">2 <i class="fa fa-flag"></i></span>
+        <span class="hk-item">Datasource Alerts</span>
+      </div>
+      <div class="col-sm-6 hk-summary-item">
+        <span class="hk-data">Chart TBD</span>
+        <span class="hk-item">Connections Used %</span>
+      </div>
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data">45 Available (of 50)</span>
+        <span class="hk-item">Available Connections (of Total)</span>
+      </div>
     </div>
+  </div>
 
 </section>
