+++ ui/console/src/main/scripts/plugins/metrics/plugins/metrics/html/app-details/detail-deployments.html
@@ -1,4 +1,4 @@
-<section id="hk-deployments" class="hk-tab-content" ng-if="vm.activeTab === 'deployments'">
+<section id="hk-deployments" class="hk-tab-content">
 
   <h3 class="pull-left">Deployment Status</h3>
   <span class="hk-update pull-right" tooltip-trigger tooltip-placement="bottom" tooltip="Last update"><i class="fa fa-clock-o"></i>{{vm.endTimeStamp | date:'mediumTime'}}</span>
