+++ ui/console/src/main/scripts/plugins/metrics/plugins/metrics/html/app-details/detail-jvm.html
@@ -1,59 +1,67 @@
-<section id="hk-jvm-metrics" class="hk-tab-content" ng-if="vm.activeTab === 'jvm-metrics'">
+<section id="hk-jvm-metrics" class="hk-tab-content">
 
-    <h3 class="pull-left">JVM Status</h3>
-    <span class="hk-update pull-right" tooltip-trigger tooltip-placement="bottom" tooltip="Last update"><i class="fa fa-clock-o"></i>{{vm.endTimeStamp | date:'mediumTime'}}</span>
-    <div class="panel panel-default hk-summary">
-        <div class="row">
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">10 <i class="fa fa-flag"></i></span>
-                <span class="hk-item">JVM Alerts</span>
-            </div>
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">1270 MB</span>
-                <span class="hk-item">Average Heap Usage</span>
-            </div>
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">5 s</span>
-                <span class="hk-item">Last Garbage Collection (Duration)</span>
-            </div>
-            <div class="col-sm-3 hk-summary-item">
-                <span class="hk-data">2000 MB</span>
-                <span class="hk-item">Maximum Heap</span>
-            </div>
-        </div>
-    </div>
+  <h3 class="pull-left">JVM Status</h3>
 
-    <div class="hk-info-top clearfix">
-        <h3 class="pull-left" tooltip-trigger tooltip-placement="top" tooltip="Expand/collapse"><button class="btn btn-link hk-trigger"><i class="fa fa-plus-square-o"></i> JVM Alerts <span>(2 of 10)</span></button></h3>
-        <span class="hk-settings pull-right"><a href="#" ng-controller="MetricsAlertController as mac" ng-click="mac.openSetup()"><i class="fa fa-cog"></i>Alert Settings</a></span>
+  <span class="hk-update pull-right" tooltip-trigger tooltip-placement="top" tooltip="Last update"><i class="fa fa-clock-o"></i>{{vm.endTimeStamp | date:'mediumTime'}}</span>
+  <div class="panel panel-default hk-summary">
+    <div class="row">
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data">{{vm.alertList.length}} <i class="fa fa-flag" ng-show="vm.alertList.length > 0"></i></span>
+        <span class="hk-item">JVM Alerts</span>
+      </div>
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data" ng-show="vm.heapUsage">{{ vm.heapUsage.avg / 1024 / 1024 | number : 0 }} MB</span>
+        <span class="hk-data" ng-hide="vm.heapUsage">n/a</span>
+        <span class="hk-item">Average Heap Usage</span>
+      </div>
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data" ng-show="vm.garbageCollector">{{ vm.garbageCollector.gc / 1024 / 1024 | number : 0 }} s</span>
+        <span class="hk-data" ng-hide="vm.garbageCollector">n/a</span>
+        <span class="hk-item">Last Garbage Collection (Duration)</span>
+      </div>
+      <div class="col-sm-3 hk-summary-item">
+        <span class="hk-data" ng-show="vm.heapUsage">{{ vm.heapUsage.max / 1024 / 1024 | number : 0 }} MB</span>
+        <span class="hk-data" ng-hide="vm.heapUsage">n/a</span>
+        <span class="hk-item">Maximum Heap</span>
+      </div>
     </div>
-    <div class="alerts-activity clearfix">
-        <div class="panel panel-default alert alert-danger">
-            <div class="hk-alert-icon"><i class="fa fa-flag"></i></div>
-            <div class="hk-date">12 Feb, 5:35 PM</div>
-            <div class="hk-event">Server <strong>down</strong> for 4 minutes (until 5:39 PM).</div>
-        </div>
-        <div class="panel panel-default alert alert-danger">
-            <div class="hk-alert-icon"><i class="fa fa-flag"></i></div>
-            <div class="hk-date">11 Feb, 3:53 PM</div>
-            <div class="hk-event"><strong>Response time</strong> was above the threshold for 7 minutes (until 4:00 PM).</div>
-        </div>
-    </div>
+  </div>
 
-    <div class="hk-info-top clearfix">
-        <h3 class="pull-left">Metrics</h3>
-        <div class="hk-update pull-right">
-            <button class="btn btn-link hk-chart-update" data-toggle="tooltip" data-placement="top" title="" data-original-title="Update chart"><i class="fa fa-refresh"></i></button>
-        </div>
+  <!-- Alerts -->
+  <div class="hk-info-top clearfix">
+    <h3 class="pull-left" data-toggle="tooltip" data-placement="top" title="" data-original-title="Expand/collapse">
+      <button class="btn btn-link hk-trigger" ng-click="showAllAlerts = !showAllAlerts"><i class="fa" ng-show="vm.alertList.length > 0" ng-class="showAllAlerts ? 'fa-minus-square-o' : 'fa-plus-square-o'"></i> JVM Alerts <span ng-show="vm.alertList.length > 0">({{showAllAlerts ? vm.alertList.length : vm.math.min(vm.alertList.length, 3)}} of {{vm.alertList.length}})</span></button>
+    </h3>
+    <span class="hk-settings pull-right"><a href="#" ng-controller="MetricsAlertController as mac" ng-click="mac.openSetup()"><i class="fa fa-cog"></i>Alert Settings</a></span>
+  </div>
+  <div class="clearfix">
+    <!-- No Alerts -->
+    <div class="panel panel-default alert alert-info" ng-show="vm.alertList.length === 0">
+      <div class="hk-alert-icon"><i class="fa fa-flag gray"></i></div>
+      <div>No alerts have been fired in this category and time range.</div>
     </div>
+    <!-- Alerts exist -->
+    <div class="panel panel-default alert alert-danger" ng-repeat="alert in vm.alertList | limitTo: (showAllAlerts ? 100000 : 3) track by $index">
+      <div class="hk-alert-icon"><i class="fa fa-flag"></i></div>
+      <div class="hk-date">{{alert.evalSets[0][0].evalTimestamp | date:'medium' }}</div>
+      <div class="hk-event"><strong>Alert Summary</strong>: alert details.</div>
+    </div>
+  </div>
 
-    <div class="panel panel-default hk-graph">
-        <h2>JVM Heap Usage</h2>
-        <div class="hk-legend pull-left">
-            <span><i class="fa fa-circle hk-primary"></i>Used</span>
-            <span><i class="fa fa-circle hk-orange"></i>Maximum</span>
-            <span><i class="fa fa-circle hk-gray"></i>Commited</span>
-        </div>
+  <div class="hk-info-top clearfix">
+    <h3 class="pull-left">Metrics</h3>
+    <div class="hk-update pull-right">
+      <button class="btn btn-link hk-chart-update" tooltip-trigger tooltip-placement="top" tooltip="Update chart"><i class="fa fa-refresh"></i></button>
     </div>
+  </div>
+
+  <div class="panel panel-default hk-graph">
+    <h2>JVM Heap Usage</h2>
+    <div class="hk-legend pull-left">
+      <span><i class="fa fa-circle hk-primary"></i>Used</span>
+      <span><i class="fa fa-circle hk-orange"></i>Maximum</span>
+      <span><i class="fa fa-circle hk-gray"></i>Commited</span>
+    </div>
+  </div>
 
 </section>
