+++ ui/console/src/main/scripts/plugins/metrics/plugins/metrics/ts/app-details/appServerJmsDetails.ts
@@ -1,0 +1,28 @@
+///
+/// Copyright 2015 Red Hat, Inc. and/or its affiliates
+/// and other contributors as indicated by the @author tags.
+///
+/// Licensed under the Apache License, Version 2.0 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///    http://www.apache.org/licenses/LICENSE-2.0
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+
+/// <reference path="../metricsPlugin.ts"/>
+/// <reference path="../alertsManager.ts"/>
+/// <reference path="../errorManager.ts"/>
+
+module HawkularMetrics {
+
+  export class AppServerJmsDetailsController {
+  }
+
+  _module.controller('HawkularMetrics.AppServerJmsDetailsController', AppServerJmsDetailsController);
+}
